Add files via upload
I am discussing complete end to end project on microsoft excel we have analysed the data using excel formulaes and created dashboard using excel graphs and charts. The dashboard is completely dynamic in nature with slicers.
This can be included as a advanced level project into your resume and will give a visibility to the recruiter that you know the skills and how to apply them practically.